+++ db/migrate/20180808164416_add_approval_sent_at_to_users.rb
@@ -1,0 +1,5 @@
+class AddApprovalSentAtToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :approval_sent_at, :datetime
+  end
+end